requirements.txt
lmdb==1.4.1
Pillow==10.2.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.4.0
tqdm==4.66.1
scipy==1.15.3
numpy==1.26.4
pandas==2.3.3
tensorboard==2.20.0
matplotlib==3.10.8
protobuf<4.0.0